Defer handling of if conditions within .irp to after the expansion
While the initial implementation in of .iflt and .irpc in
867337f38d did work for each of the constructs on its own, it
didn't work as intended for
libavcodec/arm/mpegaudiodsp_fixed_armv6.S, which used these two
constructs combined, which kind of was what that commit advertised.

This commit defers handling of all .if constructs within .irp
blocks and tries to handle them at the end, after the .irp has
been expanded.

I'm not sure if this interacts correctly with macros and other
things, but at least it seems to handle mpegaudiodsp_fixed_armv6.S.