Fix handling of thumb symbols
The RE for matching apple- was typoed in ea29a0e6c. Additionally,
reverse the order of the two checks, since we use $3 and $4 to refer
to parts of one of the REs. Prior to ea29a0e6c the as_type check
was a plain string comparison, not a RE.