Force branches to local labels to be wide instructions in armasm
A branch instruction in thumb mode might either be narrow or
wide, depending on the offset to the branch location.

armasm seems to have a bug when calculating the difference
between two symbols, where there's such a branch inbetween.
If the branch immediately seems to require a wide instruction,
the symbol difference is calculated properly, but if the
branch is made wide as a consequence from another branch
being made wide, the symbol difference isn't calculated
properly.

The bug has been reported upstream:
https://connect.microsoft.com/VisualStudio/feedback/details/839866