remove all lines starting with '#'
'#' at the start of the line indicates on all archs the start of a
comment or a preprocessor directive. This is after preprocessing so
removing the lines as comment is safe. Line number comments are stripped
by this too since gas-preprocessor.pl ignores line numbers. So it
removes just misleading debug information.

Incidentally fixes gas detection in libav's configure with Xcode on
aarch64. The temporary filename contains two '/'. The filename is added
as line number comment during preprocessing. The comment removal in
gas-preprocessor detects it on aarch64. Clang fails since the line
number comment has an unterminated string constant.