Undefine platform specific predefined defines when preprocessing C source for armasm
Since plain 'cpp' is used for preprocessing, it inherits whatever
platform specific predefined defines.

Previously this was set on the libav configure line, such as
--as='gas-preprocessor.pl -as-type armasm -- armasm -nologo -U__ELF__ -ignore 4509',
but if gas-preprocessor is to be enabled automatically, one shouldn't
need to pass -U__ELF__ via the --as flag. Also, if the option
is passed there, armasm detection fails, since "armasm -U__ELF__"
fails, since -U isn't a valid armasm option (gas-preprocessor
filters it out and only passes it to cpp).

This does break the currently unsupported setup of using the
original ARM armasm (contrary to the MS version) for building
assembly for linux though, where the __ELF__ define should be
kept. But there's not much need to support that use case, since
GNU as can be used just fine along with armcc.